+++ system/include/emscripten/html5.h
@@ -24,10 +24,8 @@ extern "C" {
 /*
  * This file defines Emscripten low-level glue bindings for interfacing with HTML5 APIs
  *
- * Documentation for the public APIs defined in this file must be updated in: 
+ * Documentation for the public APIs defined in this file must be updated in:
  *    site/source/docs/api_reference/html5.h.rst
- * A prebuilt local version of the documentation is available at: 
- *    site/build/text/docs/api_reference/html5.h.txt
  * You can also build docs locally as HTML or other formats in site/
  * An online HTML version (which may be of a different version of Emscripten)
  *    is up at http://kripken.github.io/emscripten-site/docs/api_reference/html5.h.html