+++ system/lib/wasmfs/syscalls.cpp
@@ -1132,8 +1132,7 @@ int __syscall_fchmodat(int dirfd, intptr_t path, int mode, ...) {
     // TODO: Test this case.
     return -EINVAL;
   }
-  // TODO: Handle AT_SYMLINK_NOFOLLOW once we traverse symlinks correctly.
-  auto parsed = path::parseFile((char*)path, dirfd);
+  auto parsed = path::getFileAt(dirfd, (char*)path, flags);
   if (auto err = parsed.getError()) {
     return err;
   }
@@ -1725,6 +1724,19 @@ int __syscall_recvfrom(int sockfd,
 int __syscall_recvmsg(
   int sockfd, intptr_t msg, int flags, int dummy, int dummy2, int dummy3) {
   return -ENOSYS;
+}
+
+int __syscall_fadvise64(int fd, uint64_t offset, uint64_t length, int advice) {
+  // Advice is currently ignored. TODO some backends might use it
+  return 0;
+}
+
+int __syscall__newselect(int nfds, intptr_t readfds_, intptr_t writefds_, intptr_t exceptfds_, intptr_t timeout_) {
+  // TODO: Implement this syscall. For now, we return an error code,
+  //       specifically ENOMEM which is valid per the docs:
+  //          ENOMEM Unable to allocate memory for internal tables
+  //          https://man7.org/linux/man-pages/man2/select.2.html
+  return -ENOMEM;
 }
 
 } // extern "C"